refactor BlockUtils and Utils for improved inventory checks and utility functions
